+++ responders/src/UI/embc-responder/src/app/feature-components/wizard/step-supports/step-supports.service.ts
@@ -233,11 +233,12 @@ export class StepSupportsService {
     };
     const referral: Referral = {
       ...support,
-      issuedToPersonName: this.supportDelivery.issuedTo
-        ? this.supportDelivery.issuedTo.lastName +
-          ',' +
-          this.supportDelivery.issuedTo.firstName
-        : this.supportDelivery.name,
+      issuedToPersonName:
+        this.supportDelivery.issuedTo !== 'Someone else'
+          ? this.supportDelivery.issuedTo.lastName +
+            ',' +
+            this.supportDelivery.issuedTo.firstName
+          : this.supportDelivery.name,
       method: SupportMethod.Referral,
       supplierAddress: this.supportDelivery.supplier.address,
       supplierId: this.supportDelivery.supplier.id,
