+++ responders/src/UI/embc-responder/src/app/feature-components/wizard/support-components/existing-support-details/existing-support-details.service.ts
@@ -190,8 +190,8 @@ export class ExistingSupportDetailsService {
       selectedSupport.subCategory === SupportSubCategory.Transportation_Taxi
     ) {
       return {
-        fromAddress: (selectedSupport as TransportationTaxiReferral).from,
-        toAddress: (selectedSupport as TransportationTaxiReferral).to
+        fromAddress: (selectedSupport as TransportationTaxiReferral).fromAddress,
+        toAddress: (selectedSupport as TransportationTaxiReferral).toAddress
       };
     } else if (
       selectedSupport.subCategory === SupportSubCategory.Transportation_Other
