+++ responders/src/UI/embc-responder/src/app/feature-components/wizard/support-components/view-supports/supports-table/supports-table.component.scss
@@ -14,3 +14,13 @@ table {
 .expired {
   color: rgba(201, 0, 0, 1);
 }
+
+th {
+  font-size: 16px !important;
+  background-color: rgba(113, 113, 113, 1);
+  color: white !important;
+}
+
+::ng-deep .arrow-color .mat-sort-header-arrow {
+  color: #ffff !important;
+}
