+++ responders/src/UI/embc-responder/src/app/feature-components/wizard/support-components/view-supports/view-supports.component.html
@@ -75,11 +75,10 @@
         <app-supports-table
           [supportList]="supportList"
           [filterTerm]="filterTerm"
+          [showLoader]="showLoader"
           (clickedRow)="openSupportDetails($event)"
         ></app-supports-table>
       </div>
     </div>
   </div>
 </div>
-
-<app-overlay-loader [showLoader]="showLoader"></app-overlay-loader>
